Refactor preferences endpoint controller logic to new file
